+++ src/main/java/domain/card/DealerCards.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import java.util.List;
