+++ src/main/java/domain/card/Drawable.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 public interface Drawable {
